consumer-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pika==1.3.2
pydantic==2.10.5
pydantic-settings==2.7.1
mysql-connector-python==9.1.0
pytest==8.3.4
